feat: arquitetura
