+++ package.json
@@ -2,7 +2,7 @@
   "name": "wbot",
   "version": "1.0.0",
   "description": "",
-  "main": "./app.js",
+  "main": "./src/app.js",
   "scripts": {
     "start": "node .",
     "dev": "nodemon .",
