+++ src/model/index.js
@@ -1,0 +1,7 @@
+const messageModel = require('./message_model');
+const messageCreateModel = require('./message_create_model');
+
+module.exports = {
+  messageModel,
+  messageCreateModel
+}
